requirements.txt
strawberry-graphql[debug-server]==0.235.2
SQLAlchemy==2.0.31
psycopg==3.1.18
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.1
asyncio==3.4.3
typer==0.12.3